+++ source/themes/simple/index.js
@@ -7,6 +7,7 @@ import SimpleAutocomplete from './SimpleAutocomplete.scss';
 import SimpleBubble from './SimpleBubble.scss';
 import SimpleButton from './SimpleButton.scss';
 import SimpleCheckbox from './SimpleCheckbox.scss';
+import SimpleFlex from './SimpleFlex.scss';
 import SimpleFormField from './SimpleFormField.scss';
 import SimpleHeader from './SimpleHeader.scss';
 import SimpleInput from './SimpleInput.scss';
@@ -30,6 +31,7 @@ export const SimpleTheme = {
   [IDENTIFIERS.BUBBLE]: SimpleBubble,
   [IDENTIFIERS.BUTTON]: SimpleButton,
   [IDENTIFIERS.CHECKBOX]: SimpleCheckbox,
+  [IDENTIFIERS.FLEX]: SimpleFlex,
   [IDENTIFIERS.FORM_FIELD]: SimpleFormField,
   [IDENTIFIERS.HEADER]: SimpleHeader,
   [IDENTIFIERS.INPUT]: SimpleInput,