+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## Features
 
+- Adds Flex and FlexItem components to layout components.
+  [PR 85](https://github.com/input-output-hk/react-polymorph/pull/85)
+
 - Adds Header component, Header skin, and adds SimpleHeader to SimpleTheme.
   [PR 84](https://github.com/input-output-hk/react-polymorph/pull/84)